requirements.txt
aiohttp==3.6.2
astunparse==1.6.3
async-timeout==3.0.1
attrs==19.3.0
braceexpand==0.1.5
cffi==1.14.1
chardet==3.0.4
colorama==0.4.3
colouredlogs==10.0.1
discord.py==1.4.1
humanfriendly==4.7
humanize==2.5.0
idna==2.10
import-expression==1.1.4
jishaku==1.19.0.194
multidict==4.7.6
pycparser==2.20
PyNaCl==1.3.0
pyreadline==2.1
six==1.15.0
wavelink==0.9.4
yarl==1.5.1
